+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - Filter out offline peers for data resolving #947
 
+### Fixed
+- Fixed transaction hash integrity between wallet and node #946
+
 ## [v2.0.1-rc] - 2020-03-17
 ### Changed
 - Instantiate CloudStorage for Genesis flow